DataFlash: Change ::bufferspace_available return type...
from 'uint16_t' to 'uint32_t' to avoid overflow (possible in at least
two implementations: Block and File).

+++ libraries/DataFlash/DataFlash_Block.h
@@ -44,7 +44,7 @@ public:
     void ShowDeviceInfo(AP_HAL::BetterStream *port);
     void ListAvailableLogs(AP_HAL::BetterStream *port);
 
-    uint16_t bufferspace_available();
+    uint32_t bufferspace_available();
 
 private:
     struct PageHeader {
